+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 #### next release (8.3.8)
 
+- Fix maximum call stack size exceeded on Math.min/max when creating Charts
+- Fix boolean flag in `MyDataTab` displaying number
 - Remove `jsx-control-statements` dependency
 - Fix WMS nested group IDs - nested groups with the same name were not being created
 - WMS `isEsri` default value will now check for case-insensitive `mapserver/wmsserver` (instead of `MapServer/WMSServer`)